:recycle: make a new data-widget file
for widget specific data methods in the new namespace.  Move the adaptive_section_slice method there, since only the widget needs it.
