+++ bu-navigation.php
@@ -57,6 +57,7 @@ require_once BU_NAV_PLUGIN_DIR . '/includes/class-tree-view.php';
 require_once BU_NAV_PLUGIN_DIR . '/includes/class-reorder.php';
 require_once BU_NAV_PLUGIN_DIR . '/includes/data-model.php';
 require_once BU_NAV_PLUGIN_DIR . '/includes/data-format.php';
+require_once BU_NAV_PLUGIN_DIR . '/includes/data-widget.php';
 
 /**
  * Convenience class to wrap loading and init functions.
